+++ Simix.SuperMidia.DeviceConnect/DevicesView.resx
