+++ Simix.SuperMidia.DeviceConnect/Program.cs
@@ -14,7 +14,7 @@ namespace RemoteDevices {
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new DevicesView());
         }
     }
 }
