+++ Simix.SuperMidia.Utils.Core/Adb/LanguageMode.cs
@@ -1,0 +1,7 @@
+﻿namespace Simix.SuperMidia.Utils.Core.Adb {
+    public enum LanguageMode {
+        Pt_BR = 0,
+        Es_ES = 1,
+        En_US = 2
+    }
+}
