+++ Simix.SuperMidia.Utils.Core/Adb/SuperUser.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.ComponentModel;
+using System.Xml.Serialization;
+
+namespace Simix.SuperMidia.Utils.Core.Adb {
+    [Serializable()]
+    [DesignerCategory("code")]
+    [XmlType(AnonymousType = true)]
+    public partial class SuperUser : AdbCommand {
+    }
+}
